# Conflicts:
#	src/test/java/org/team1/keyduck/keyboard/service/KeyboardServiceTest.java
#	src/test/java/org/team1/keyduck/testdata/TestData.java

+++ src/main/java/org/team1/keyduck/auction/dto/response/AuctionCreateResponseDto.java
@@ -21,7 +21,7 @@ public class AuctionCreateResponseDto {
 
     private final Long currentPrice;
 
-    private final int biddingUnit;
+    private final Long biddingUnit;
 
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private final LocalDateTime auctionStartDate;
@@ -30,7 +30,7 @@ public class AuctionCreateResponseDto {
     private final LocalDateTime auctionEndDate;
 
     private AuctionCreateResponseDto(Long auctionId, AuctionKeyboardDto keyboard, String title,
-            Long startPrice, Long immediatePurchasePrice, Long currentPrice, int biddingUnit,
+            Long startPrice, Long immediatePurchasePrice, Long currentPrice, Long biddingUnit,
             LocalDateTime auctionStartDate, LocalDateTime auctionEndDate) {
         this.auctionId = auctionId;
         this.keyboard = keyboard;
